Django/Films_API/requirements.txt
django==4.2.11
djangorestframework==3.14.0
drf-spectacular===0.27.1
whitenoise===6.6.0
django-filter==24.2
django-cors-headers===4.3.1